skillprotect msc/pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "python-template"
version = "0.1.0"
[tool.poetry.dependencies]
Augmentor = "^0.2.9"
DisCapTy = "^1.0.2"
Flask = "^2.0.3"
Strip = "^0.0.1-alpha.14"
asyncpg = "^0.25.0"
attribute = "^0.1.4"
button = "^0.0.3"
buttons = "^0.1.9"
config = "^0.5.1"
db = "^0.1.1"
db2 = "^0.1.1"
discord = "^1.7.3"
discord-ext-slash = "^0.9.2"
discord-ui = "^5.1.6"
"dislash.py" = "^1.4.9"
embed = "^0.1.0"
handler = "^1.3.0"
humanfriendly = "^10.0"
kwargs = "^1.0.1"
mkdocs-user-defined-values = "^0.0.5"
numpy = "^1.22.2"
pymongo = "^4.0.2"
python = "^3.8"
python-dotenv = "^0.19.2"
pytz = "^2021.3"
reactionmenu = "^3.0.1"
requests = "^2.27.1"
rmenulegacy = "^2.0.3"
session3 = "^3.0.post0"
srv = "^1.1.0"
ui = "^0.1.4"
utils = "^1.0.1"
validators = "^0.18.2"
[tool.poetry.dev-dependencies]
